zion: x86: add a simple arch specific kmalloc
This patch adds a simple kmalloc implementation for dynamically
allocating memory. Note that this is marked unsafe as going
beyond the mapped 4MB of area will lead to a page fault. This
is only really meant to be used in the initial phase of the
boot process before a proper allocation scheme is set up.

Signed-off-by: Manohar Vanga <mvanga@mpi-sws.org>